chromium/third_party/blink/web_tests/external/wpt/streams/piping/throwing-options.any-expected.txt

This is a testharness.js-based test.
FAIL pipeTo should stop after getting preventAbort throws assert_array_equals: options should be touched in the right order lengths differ, expected array ["preventAbort"] length 1, got ["preventClose", "preventAbort"] length 2
FAIL pipeThrough should stop after getting preventAbort throws assert_array_equals: options should be touched in the right order lengths differ, expected array ["preventAbort"] length 1, got ["preventClose", "preventAbort"] length 2
FAIL pipeTo should stop after getting preventCancel throws assert_array_equals: options should be touched in the right order lengths differ, expected array ["preventAbort", "preventCancel"] length 2, got ["preventClose", "preventAbort", "preventCancel"] length 3
FAIL pipeThrough should stop after getting preventCancel throws assert_array_equals: options should be touched in the right order lengths differ, expected array ["preventAbort", "preventCancel"] length 2, got ["preventClose", "preventAbort", "preventCancel"] length 3
FAIL pipeTo should stop after getting preventClose throws assert_array_equals: options should be touched in the right order lengths differ, expected array ["preventAbort", "preventCancel", "preventClose"] length 3, got ["preventClose"] length 1
FAIL pipeThrough should stop after getting preventClose throws assert_array_equals: options should be touched in the right order lengths differ, expected array ["preventAbort", "preventCancel", "preventClose"] length 3, got ["preventClose"] length 1
FAIL pipeTo should stop after getting signal throws assert_array_equals: options should be touched in the right order expected property 0 to be "preventAbort" but got "preventClose" (expected array ["preventAbort", "preventCancel", "preventClose", "signal"] got ["preventClose", "preventAbort", "preventCancel", "signal"])
FAIL pipeThrough should stop after getting signal throws assert_array_equals: options should be touched in the right order expected property 0 to be "preventAbort" but got "preventClose" (expected array ["preventAbort", "preventCancel", "preventClose", "signal"] got ["preventClose", "preventAbort", "preventCancel", "signal"])
Harness: the test ran to completion.
